Добавил форму для комментариев, вынес формы в отдельную директорию.
